backend/requirements.txt
# GabomaGPT — Dépendances Backend · SmartANDJ AI Technologies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Base de données
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# IA et LLM
groq==0.13.0
openai==1.58.1
tavily-python==0.5.0

# RAG
chromadb==0.5.23
rank-bm25==0.2.2
sentence-transformers==3.3.1

# Cache et sessions
redis==5.2.1

# Agent Black Panther
playwright==1.49.1

# Utilitaires
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0
